refactor hotkeys to hotkeys.py
